pyproject.toml
[tool.poetry]
name = "bond-ai"
version = "0.1.1"
description = ""
authors = ["jmcarnahan <jmcarny@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^8.3.4"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
streamlit-google-auth = "^1.1.8"
ipython = "^8.31.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
